requirements.txt
celery
django<3
django-celery-beat
django-debug-toolbar
django-extensions
django-statsd-mozilla
flower
gunicorn
mysqlclient
redis<4
requests
statsd
tornado<5
Werkzeug
